Refactor the way decorators are handled.
This is to be able to stack them in a nice way, without having
to go many times in the different verification processes.

This commit changes the way the verification are made.
Internal decorators are not used anymore so it is more readable
and easier to understand.

Documentation is now generated only by the sphinx directive, not
before.