Change the internal implementation of accept.
Previously, Cornice was doing its own check rather than using the already
defined mechanisms in pyramid. Now, cornice rely on pyramid view predicates.

One thing was not handled according to the http specification: when a 406 Not
Acceptable is returned, a list of acceptable Content-Type values should be
returned as well. It's now the case.

This is being handled by a custom 404 generator, which relies on some pyramid
internal concepts (for instance, a PredicateMismatch exception is thrown in
case none of the predicates are met, and it is thrown as a 404 error. We
intercept it and return a 406 instead)