Adds a filters management and rework validators.
Validators can be registered application-wide, service-wide or only for a
method. It is also possible to exclude some previously included
filters/validators, by using the exclude parameter (see the doc).

validator is now named validator*s*. (but the former is still valid, even if
it's not the recommended way to go)