Fix the test on dropped param in colander schema.
This test used to test nothing as the body defaulted to {} as the mocked request
didn't find any way to deserialize (content_type was not set and registry was
an dict where it should be an object.
The test just checked that the field was not present.

Now we check that 'bar' is present (to avoid false test pass).