Remove tests of unsupported "default" values for colander schemas
Before we try to make colander work, and handle "default" values in
incompatible with colander way. We thread as default value of "default"
atribut. But colander treat as default value of "missing" attribut for
deserialize operation.

Because we don't do colander job any more, we can remove tests used to test
incompatible behaviour.