Drops simplejson as requirement
To achieve that, a new setting called ``cornice_default_renderer`` is added,
set by default to ``'simplejson'``.
If the ``cornice_default_renderer`` is simplejson, the serializer will be patched
the same way it used to be before with ``simplejson.dumps(..., use_decimal=True)``
to keep retrocompatibility.
Otherwise, the default serializer of pyramid will be used, unless overridden.